fix(M8-audit): missing needs-qa→needs-human transition; move retry-counter to core/retry/
