+++ core/state-machine/transitions.ts
@@ -20,8 +20,8 @@ const TRANSITIONS: Readonly<Record<StateName, readonly StateName[]>> = {
   'factory:investigation-complete': ['factory:dev-ready'],
   'factory:dev-ready': ['factory:in-progress'],
   'factory:in-progress': ['factory:needs-qa'],
-  'factory:needs-qa': ['factory:qa-failed', 'factory:needs-review'],
-  'factory:qa-failed': ['factory:needs-fix'],
+  'factory:needs-qa': ['factory:qa-failed', 'factory:needs-review', 'factory:needs-human'],
+  'factory:qa-failed': ['factory:needs-fix', 'factory:needs-human'],
   // needs-review → rejected covers the "human explicitly cancelled" case from section 9.2
   'factory:needs-review': [
     'factory:needs-fix',
